refactor: scaffold applicant library and website binary
