+++ Cargo.toml
@@ -1,9 +1,7 @@
-[package]
-name = "Battalion"
-version = "0.1.0"
-edition = "2024"
+[workspace]
+members = ["website", "applicant"]
 
-[dependencies]
+[workspace.dependencies]
 axum = "0.8.4"
 serde = { version = "1.0.219", features = ["derive"] }
 serde_json = "1.0.140"
