+++ applicant/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "applicant"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+axum = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+surrealdb = { workspace = true }
+tokio = { workspace = true }
+tera = { workspace = true }
+schemars = { workspace = true }
+url = { workspace = true }
