+++ applicant/src/routes.rs
@@ -1,0 +1,14 @@
+use std::sync::Arc;
+use axum::{Router, routing::get};
+use crate::{AppState, handlers};
+
+pub fn routes() -> Router<Arc<AppState>> {
+    Router::new()
+        .route("/apply", get(handlers::show_form).post(handlers::submit_form))
+        .route(
+            "/apply/{id}",
+            get(handlers::fetch_form)
+                .put(handlers::update_form)
+                .delete(handlers::delete_form),
+        )
+}
