+++ applicant/templates/applicant_list.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="utf-8">
+    <title>Applicant List</title>
+</head>
+<body>
+    <h1>Applicant List</h1>
+    <!-- TODO: list applicants -->
+</body>
+</html>
