+++ website/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "website"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+axum = { workspace = true }
+tokio = { workspace = true }
+tower-http = { workspace = true }
+tera = { workspace = true }
+surrealdb = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+schemars = { workspace = true }
+url = { workspace = true }
+applicant = { path = "../applicant" }
